# frontend/Agricultural_Products_Market_admin/default_OpenAPI.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
...
  "paths": {
...
    "/admin/order/status": {
      "put": {
        "tags": [
          "管理员订单管理"
        ],
        "summary": "修改订单状态",
        "description": "更新订单状态，如将已支付订单标记为已发货",
        "operationId": "updateOrderStatus",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOrderStatusDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
    "/admin/category/{id}": {
      "put": {
        "tags": [
          "管理员分类管理"
        ],
        "summary": "修改分类",
        "description": "管理员修改农产品分类",
        "operationId": "updateCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduceCategoryDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "管理员分类管理"
        ],
        "summary": "删除分类",
        "description": "管理员删除农产品分类",
        "operationId": "deleteCategory",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/category": {
      "post": {
        "tags": [
          "管理员分类管理"
        ],
        "summary": "添加分类",
        "description": "管理员添加农产品分类",
        "operationId": "addCategory",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProduceCategoryDTO"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultString"
                }
              }
            }
          }
        }
      }
    },
...
    "/order/{id}": {
      "get": {
        "tags": [
          "订单管理"
        ],
        "summary": "查看订单详情",
        "description": "获取指定订单的详细信息，包括订单项",
        "operationId": "getOrderDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultOrderVO"
                }
              }
            }
          }
        }
      }
    },
...
    "/category/{id}": {
      "get": {
        "tags": [
          "农产品分类管理"
        ],
        "summary": "获取农产品分类详情",
        "description": "根据分类ID获取农产品分类详情",
        "operationId": "getCategoryDetail",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultProduceCategory"
                }
              }
            }
          }
        }
      }
    },
...
    "/admin/order": {
      "get": {
        "tags": [
          "管理员订单管理"
        ],
        "summary": "分页展示所有用户订单",
        "description": "分页查询所有用户订单，支持按下单日期和支付状态筛选",
        "operationId": "getOrderList_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "payStatus",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ResultIPageOrderInfo"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ResultString": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "type": "string"
          }
        }
      },
...
      "UpdateOrderStatusDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "orderStatus": {
            "type": "string"
          }
        }
      },
      "ProduceCategoryDTO": {
        "type": "object",
        "properties": {
          "categoryName": {
            "type": "string",
            "description": "分类名称"
          }
        },
        "description": "农产品分类DTO"
      },
...
      "OrderInfo": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "订单ID",
            "format": "int64"
          },
          "orderNo": {
            "type": "string",
            "description": "订单编号"
          },
          "userId": {
            "type": "integer",
            "description": "用户ID",
            "format": "int64"
          },
          "addressId": {
            "type": "integer",
            "description": "地址ID",
            "format": "int64"
          },
          "totalAmount": {
            "type": "number",
            "description": "订单总金额"
          },
          "payStatus": {
            "type": "integer",
            "description": "支付状态：0-未支付，1-已支付",
            "format": "int32"
          },
          "orderStatus": {
            "type": "string",
            "description": "订单状态：待发货/已发货/已完成"
          },
          "payTime": {
            "type": "string",
            "description": "支付时间",
            "format": "date-time"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          },
          "updateTime": {
            "type": "string",
            "description": "更新时间",
            "format": "date-time"
          }
        },
        "description": "订单表"
      },
...
      "OrderItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "订单项ID",
            "format": "int64"
          },
          "orderId": {
            "type": "integer",
            "description": "订单ID",
            "format": "int64"
          },
          "produceId": {
            "type": "integer",
            "description": "农产品ID",
            "format": "int64"
          },
          "produceName": {
            "type": "string",
            "description": "农产品名称"
          },
          "price": {
            "type": "number",
            "description": "购买单价"
          },
          "quantity": {
            "type": "integer",
            "description": "购买数量",
            "format": "int32"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          }
        },
        "description": "订单项明细表"
      },
      "OrderVO": {
        "type": "object",
        "properties": {
          "orderInfo": {
            "$ref": "#/components/schemas/OrderInfo"
          },
          "orderItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderItem"
            }
          },
          "recipient": {
            "type": "string"
          },
          "phone": {
            "type": "string"
          },
          "address": {
            "type": "string"
          }
        }
      },
      "ResultOrderVO": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/OrderVO"
          }
        }
      },
      "ProduceCategory": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "description": "分类ID",
            "format": "int64"
          },
          "categoryName": {
            "type": "string",
            "description": "分类名称"
          },
          "createTime": {
            "type": "string",
            "description": "创建时间",
            "format": "date-time"
          }
        },
        "description": "农产品分类表"
      },
...
      "ResultProduceCategory": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/ProduceCategory"
          }
        }
      },
...
      "IPageOrderInfo": {
        "type": "object",
        "properties": {
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "pages": {
            "type": "integer",
            "format": "int64",
            "deprecated": true
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OrderInfo"
            }
          },
          "current": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "ResultIPageOrderInfo": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "msg": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/IPageOrderInfo"
          }
        }
      },
...
    }
  },
...
}